commit 30/04/2020 ParticiperBien

+++ src/main/java/com/proxym/pfe/gestionAssociationBack/biens/entities/ParticiperBien.java
@@ -12,7 +12,7 @@ public class ParticiperBien {
     @Id
     private Long idP;
     @Column(name = "qteDonnee")
-    private Long qteDonnee;
+    private int qteDonnee;
     @ManyToOne
     private User user;
 }
